Try fix unlink() on windows

+++ Lib/test/support/__init__.py
@@ -376,12 +376,24 @@ if sys.platform.startswith("win"):
                       RuntimeWarning, stacklevel=4)
 
     def _unlink(filename):
+        # XXX RUSTPYTHON: on ci, unlink() raises PermissionError when target doesn't exist.
+        # Might also happen locally, but not sure
+        if not os.path.exists(filename):
+            return
         _waitfor(os.unlink, filename)
 
     def _rmdir(dirname):
+        # XXX RUSTPYTHON: on ci, unlink() raises PermissionError when target doesn't exist.
+        # Might also happen locally, but not sure
+        if not os.path.exists(dirname):
+            return
         _waitfor(os.rmdir, dirname)
 
     def _rmtree(path):
+        # XXX RUSTPYTHON: on ci, unlink() raises PermissionError when target doesn't exist.
+        # Might also happen locally, but not sure
+        if not os.path.exists(path):
+            return
         def _rmtree_inner(path):
             for name in _force_run(path, os.listdir, path):
                 fullname = os.path.join(path, name)
@@ -1051,7 +1063,15 @@ def temp_dir(path=None, quiet=False):
         # In case the process forks, let only the parent remove the
         # directory. The child has a different process id. (bpo-30028)
         if dir_created and pid == os.getpid():
-            rmtree(path)
+            try:
+                rmtree(path)
+            except OSError as exc:
+                # XXX RUSTPYTHON: added quiet check here, not sure why rmtree fails on windows
+                if not quiet:
+                    raise
+                warnings.warn(f'unable to remove temporary'
+                              f'directory {path!r}: {exc}',
+                              RuntimeWarning, stacklevel=3)
 
 @contextlib.contextmanager
 def change_cwd(path, quiet=False):
